code/server/app/requirements.txt
flask
flask-jsonpify
flask-restful
Flask
watchdog
cassandra-driver
argparse
uuid
multiprocess
paho-mqtt